app-gateway-pl/pyproject.toml
[project]
name = "aurabc-appgw-validation"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "neo4j",
    "python-dotenv",
    "requests",
]
